+++ src/Products/MeetingLalouviere/profiles/zlalouviere/import_data.py
@@ -154,7 +154,6 @@ collegeMeeting.usedItemAttributes = (
     u"observations",
     u"manuallyLinkedItems",
     u"textCheckList",
-    u"neededFollowUp",
     u"providedFollowUp",
 )